+++ R/fct_derived_traits.R
@@ -46,7 +46,7 @@ generate_formula <- function(cluster, cluster_ref_df, target_trait) {
   # Divide by the sum of all complex-type glycans if necessary
   # Or by the sum of all oligomannose glycans
   if (target_trait %in% c("fucosylation", "bisection", "galactosylation", "sialylation", 
-                          "sialylation_per_galactose", "mono_antennary", "alpha_galactosylation")) {
+                          "mono_antennary", "alpha_galactosylation")) {
     complex_types_df <- cluster_ref_df %>% 
       dplyr::select(glycan, complex) %>% 
       dplyr::filter(complex == 1)
@@ -96,7 +96,7 @@ match_human_IgG_traits <- function(human_traits_ui_input) {
     "Bisection of complex-type glycans" = "bisection",
     "Galactosylation per antenna of complex-type glycans" = "galactosylation",
     "Sialylation per antenna of complex-type glycans" = "sialylation",
-    "Sialylation per galactose of complex-type glycans" = "sialylation_per_galactose",
+    "Sialylation per galactose of complex-type glycans (calculated as [Sialylation per antenna] / [Galactosylation per antenna] \u00D7 100%)" = "sialylation_per_galactose",
     "Percentage of monoantennary complex-type glycans" = "mono_antennary",
     "Percentage of hybrid-type glycans" = "hybrid",
     "Percentage of oligomannose-type glycans" = "oligomannose_relative",
@@ -159,6 +159,9 @@ match_mouse_IgG_traits <- function(mouse_traits_ui_input) {
 #' @param reference Reference file for traits, e.g. human_IgG_N_ref.
 #' 
 create_formula_list <- function(normalized_data, chosen_traits, chosen_clusters, reference) {
+  # Remove sialylation per galactose from the vector with chosen traits
+  # Sialylation per galactose is calculated manually in the server part.
+  chosen_traits <- chosen_traits[chosen_traits != "sialylation_per_galactose"]
   # Create an empty vector to store possible analytes with unknown glycan compositions
   unknown_glycans <- c()
   # Initiate an empty list